EMP-746: Refactored service tests
