+++ equinity-historical-data/src/test/java/uk/gov/justice/laa/crime/equinity/historicaldata/service/report/Crm4PeriodicalReportServiceTest.java
@@ -11,8 +11,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import uk.gov.justice.laa.crime.equinity.historicaldata.model.report.Crm5UpperLimitReportModel;
-import uk.gov.justice.laa.crime.equinity.historicaldata.repository.report.Crm5UpperLimitReportRepository;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.report.Crm4PeriodicalReportModel;
+import uk.gov.justice.laa.crime.equinity.historicaldata.repository.report.Crm4PeriodicalReportRepository;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -24,37 +24,29 @@ import static org.mockito.Mockito.when;
 @SpringBootTest
 @ExtendWith(SoftAssertionsExtension.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class Crm5UpperLimitReportServiceMockTest {
+class Crm4PeriodicalReportServiceTest {
     @InjectSoftAssertions
     private SoftAssertions softly;
 
     @MockBean
-    Crm5UpperLimitReportRepository reportRepository;
+    Crm4PeriodicalReportRepository reportRepository;
 
     @Autowired
-    Crm5UpperLimitReportService reportService;
+    Crm4PeriodicalReportService reportService;
 
     @BeforeAll
     void setup() {
-        List< Crm5UpperLimitReportModel> expectedResponse = new ArrayList<>();
-        Crm5UpperLimitReportModel report = new Crm5UpperLimitReportModel(
-                5001600L,
-                "0D182J",
-                "ABELS",
-                "Joe modo",
-                "031022/777",
-                238,
-                239,
-                LocalDate.of(2023, 3, 16),
-                LocalDate.of(2023, 3, 16),
-                "POOL-Mock",
-                "Appeal",
-                "Advice"
-        );
+        List<Crm4PeriodicalReportModel> expectedResponse = new ArrayList<>();
+        Crm4PeriodicalReportModel report = Crm4PeriodicalReportModel.builder()
+                .clientUfn("031022/777").usn(5001600L).providerAccount("0D182J").firmName("ABELS").clientName("Joe modo")
+                .repOrderNumber("78543657").maatId("").prisonLaw("No").receivedDate(LocalDate.of(2023, 3, 16))
+                .decisionDate(LocalDate.of(2023, 3, 16)).decisionResult("Grant").expenditureType("a Psychiatrist")
+                .expertName("tyjtjtjt").quantity(4.0).rate(50.0).unit("Hour(s)").totalCost(200.0).additionalExpenditure(0.0)
+                .totalAuthority(200.0).totalGranted(200.0).grantingCaseworker("Sym-G").build();
+
         expectedResponse.add(report);
 
-        when(reportRepository.getReport(any(), any()))
-                .thenReturn(expectedResponse);
+        when(reportRepository.getReport("2010-02-01", "2024-06-01")).thenReturn(expectedResponse);
     }
 
     @AfterAll
@@ -67,7 +59,7 @@ class Crm5UpperLimitReportServiceMockTest {
         String results = reportService.getReport("2010-02-01", "2024-06-01");
 
         softly.assertThat(results).isNotEmpty();
-        softly.assertThat(results).startsWith(Crm5UpperLimitReportModel.CSV_HEADER);
-        softly.assertThat(results).contains("5001600,");
+        softly.assertThat(results).startsWith(Crm4PeriodicalReportModel.CSV_HEADER);
+        softly.assertThat(results).contains(",5001600,");
     }
 }
